refactor(piece): 코드 중복 제거
