+++ src/main/java/chess/domain/piece/King.java
@@ -4,7 +4,7 @@ import chess.domain.board.position.Position;
 import java.util.List;
 import java.util.function.BiPredicate;
 
-public class King extends Piece {
+public class King extends UnpromotablePiece {
 
     private static final String SYMBOL = "k";
     private static final double SCORE = 0;
@@ -38,16 +38,6 @@ public class King extends Piece {
     @Override
     public boolean isPawn() {
         return false;
-    }
-
-    @Override
-    public boolean canPromote() {
-        return false;
-    }
-
-    @Override
-    public Piece promote(final String promotionType) {
-        throw new IllegalStateException("King은 Promotion 할 수 없습니다.");
     }
 
     @Override
