+++ src/main/java/chess/domain/piece/Knight.java
@@ -4,7 +4,7 @@ import chess.domain.board.position.Position;
 import java.util.List;
 import java.util.function.BiPredicate;
 
-public class Knight extends Piece {
+public class Knight extends UnpromotablePiece {
 
     static final String SYMBOL = "n";
     private static final double SCORE = 2.5;
@@ -39,16 +39,6 @@ public class Knight extends Piece {
     @Override
     public boolean isPawn() {
         return false;
-    }
-
-    @Override
-    public boolean canPromote() {
-        return false;
-    }
-
-    @Override
-    public Piece promote(final String promotionType) {
-        throw new IllegalStateException("Knight는 Promotion 할 수 없습니다.");
     }
 
     @Override
