+++ src/main/java/chess/domain/piece/Queen.java
@@ -4,7 +4,7 @@ import chess.domain.board.position.Position;
 import java.util.List;
 import java.util.function.BiPredicate;
 
-public class Queen extends Piece {
+public class Queen extends UnpromotablePiece {
 
     static final String SYMBOL = "q";
     private static final double SCORE = 9;
@@ -40,16 +40,6 @@ public class Queen extends Piece {
     @Override
     public boolean isPawn() {
         return false;
-    }
-
-    @Override
-    public boolean canPromote() {
-        return false;
-    }
-
-    @Override
-    public Piece promote(final String promotionType) {
-        throw new IllegalStateException("Queen은 Promotion 할 수 없습니다.");
     }
 
     @Override
